sync

+++ js/opack.js
@@ -39,9 +39,10 @@ var verbs = {
 	"update": {
 		"help"		  : "Updates a package",
 		"optionshelp" : [ "The install/erase verbs options are valid also for update.",
-						  "'-all'  : Tries to update all packages locally installed",
-						  "'-cred' : Provide authentication credentials (e.g. user:pass)",
-		                  "'-erase': When updating delete the package frist"]
+				"'-all'       : Tries to update all packages locally installed",
+				"'-cred'      : Provide authentication credentials (e.g. user:pass)",
+				"'-noerase'   : When updating don't delete the package first",
+			  "'-erasefolder: Erase previous version folder while updating to a new version"]
 	},
 	"exec": {
 		"help"		  : "Executes code from an installed package",
@@ -628,8 +629,9 @@ function execHTTPWithCred(aURL, aRequestType, aIn, aRequestMap, isBytes, aTimeou
 
 // Find OpenAF she-bang
 function getOpenAFSB() {
-  var os = java.lang.System.getProperty("os.name") + "";
-  var currentClassPath = java.lang.management.ManagementFactory.getRuntimeMXBean().getClassPath() + "";
+	var os = String(java.lang.System.getProperty("os.name"));
+	var currentClassPath = ow.format.getClasspath();
+  //var currentClassPath = java.lang.management.ManagementFactory.getRuntimeMXBean().getClassPath() + "";
 
   var openafsb;
   if (os.match(/Windows/)) {
@@ -653,16 +655,14 @@ function getOpenAFSB() {
 }
 
 // Find OpenAF jar
-function getOpenAFJar() {
-  	var currentClassPath = java.lang.management.ManagementFactory.getRuntimeMXBean().getClassPath() + "";
-
-  	return currentClassPath;
-}
+/*function getOpenAFJar() {
+  return String(ow.format.getClasspath());
+}*/
 
 // Find OpenAF
 function getOpenAF() {
-  var os = java.lang.System.getProperty("os.name") + "";
-  var currentClassPath = java.lang.management.ManagementFactory.getRuntimeMXBean().getClassPath() + "";
+	var os = ow.format.getOS();
+  var currentClassPath = ow.format.getClasspath();
 
   var openaf;
   if (os.match(/Windows/)) {
@@ -833,24 +833,24 @@ function getPackage(packPath) {
 
 // INFO
 function __opack_info(args) {
-	if(isUndefined(args[0]) || args[0] == "") return;
+	if(isUnDef(args[0]) || args[0] == "") return;
 
 	var packag = getPackage(args[0]);
 	var remote = false;
 
-	if (isUndefined(packag) || isUndefined(packag["name"])) {
+	if (isUnDef(packag) || isUnDef(packag["name"])) {
 		packag = findLocalDBByName(args[0]);
 
-		if (isUndefined(packag) || isUndefined(packag["name"])) {
+		if (isUnDef(packag) || isUnDef(packag["name"])) {
 			var packs = getRemoteDB();
 			var packFound = findCaseInsensitive(packs, args[0]);
-			if (isUndefined(packFound)) {
+			if (isUnDef(packFound)) {
 				logErr("No entry for '" + args[0] + "' on remote OPack database.");
 				logErr("Package not found.");
 				return;
 			} else {
 				packag = getPackage(packFound.repository.url);
-				if (isUndefined(packag) && !(isUndefined(packFound.repository.backupurl))) {
+				if (isUnDef(packag) && !(isUnDef(packFound.repository.backupurl))) {
 					packag = getPackage(packFound.repository.backupurl)
 				}
 				args[0] = packag.repository.url;
@@ -955,10 +955,10 @@ function checkVersion(packag, force) {
 
 // INSTALL
 function install(args) {
-	if (isUndefined(args[0]) || args[0].length <= 0)
+	if (isUnDef(args[0]) || args[0].length <= 0)
 		return;
 
-	if (!isUndefined(args[0]) && args[0].toUpperCase() == 'OPENAF') {
+	if (!isUnDef(args[0]) && args[0].toUpperCase() == 'OPENAF') {
 		logErr("Can't install OpenAF. Please install manually or using another package manager (e.g. RPM). To update execute 'openaf --update'.");
 		return;
 	}
@@ -969,8 +969,8 @@ function install(args) {
     var foundOutput = false;
     var force = false;
     var foundRepo = false;
-	var foundArg = false;
-	var foundCred = false;
+	  var foundArg = false;
+	  var foundCred = false;
     var justCopy = false;
     var useunzip = false;
     var nohash = false;
@@ -1008,19 +1008,19 @@ function install(args) {
     	if (args[i] == "-arg") foundArg = true;
     	if (args[i] == "-justcopy") justCopy = true;
     	if (args[i] == "-noverify") nohash = true;
-		if (args[i] == "-useunzip") useunzip = true;
-		if (args[i] == "-cred") foundCred = true;
+		  if (args[i] == "-useunzip") useunzip = true;
+		  if (args[i] == "-cred") foundCred = true;
     }
 	var packag = getPackage(args[0]);
 	if (isUnDef(output)) output = getOpenAFPath() + "/" + packag.name;
 
-	if (isUndefined(packag.name)) {
+	if (isUnDef(packag.name) && packag.__filelocation != "opackurl") {
 		//logErr("Couldn't find package on location " + args[0]);
 
 		log("Checking remote OPack database");
 		var packs = getRemoteDB();
 		var packFound = findCaseInsensitive(packs, args[0]);
-		if (isUndefined(packFound)) {
+		if (isUnDef(packFound)) {
 			logErr("No entry for '" + args[0] + "' on remote OPack database.");
 			return;
 		} else {
@@ -1034,22 +1034,8 @@ function install(args) {
 	if (checkVersion(packag, force) || justCopy) {
 		log((justCopy ? "COPYING" : "INSTALLING") + " -- " + packag.name + " version " + packag.version);
 	} else {
-		log("Checking remote OPack database");
-		var packs = getRemoteDB();
-		var packFound = findCaseInsensitive(packs, args[0]);
-		if (isUndefined(packFound)) {
-			logErr("No entry for '" + args[0] + "' on remote OPack database.");
-			return;
-		} else {
-			packag = getPackage(packFound.repository.url);
-			if (!forceOutput) output = getOpenAFPath() + "/" + packag.name;
-			args[0] = packag.repository.url;
-		}
-		if (checkVersion(packag, force)) {
-			log("UPDATING -- " + packag.name + " version " + packag.version);
-		} else {
-			return;
-		}
+		log("No need to install/update " + args[0]);
+		return;
 	}
 
 	// Verify deps
@@ -1069,7 +1055,7 @@ function install(args) {
 				log("Checking remote OPack database");
 				var packs = getRemoteDB();
 				var packFound = findCaseInsensitive(packs, i);
-				if (isUndefined(packFound)) {
+				if (isUnDef(packFound)) {
 					logErr("No entry for '" + i + "' on remote OPack database.");
 					return;
 				}
@@ -1096,7 +1082,7 @@ function install(args) {
 		}
 
 	outputPath = output;
-    if (!isUndefined(packag.scripts.preinstall) && !justCopy) runScript(packag.scripts.preinstall);
+    if (!isUnDef(packag.scripts.preinstall) && !justCopy) runScript(packag.scripts.preinstall);
 
 	switch(packag.__filelocation) {
 		case "url":
@@ -1367,7 +1353,7 @@ function __opack_script(args, isDaemon, isJob) {
 
 // UPDATE
 function update(args) {
-	if (!isUndefined(args[0]) && args[0].toUpperCase() == 'OPENAF') {
+	if (!isUnDef(args[0]) && args[0].toUpperCase() == 'OPENAF') {
 		logErr("Please use 'openaf --update' to update OpenAF");
 		return;
 	}
@@ -1377,18 +1363,19 @@ function update(args) {
 	var foundArg = false;
 	var foundCred = false;
 	var all = false;
-	var erase = false;
+	var ferase = true;
+	var derase = true;
 
-    for(let i in args) {
-    	if (foundRepo) {
-    		if (__opackCentral.indexOf(args[i]) < 0) __opackCentral.unshift(args[i]);
-    		foundRepo = false;
-    	}
+	for(let i in args) {
+		if (foundRepo) {
+			if (__opackCentral.indexOf(args[i]) < 0) __opackCentral.unshift(args[i]);
+			foundRepo = false;
+		}
 
-    	if (foundArg) {
-    		arg = args[i];
-    		foundArg = false;
-    	}
+		if (foundArg) {
+			arg = args[i];
+			foundArg = false;
+		}
 
 		if (foundCred) {
 			var cred = args[i];
@@ -1396,41 +1383,42 @@ function update(args) {
 			foundCred = false;
 		}
 
-    	if (args[i] == "-arg") foundArg = true;
-    	if (args[i] == "-force") force = true;
-    	if (args[i] == "-repo") foundRepo = true;
-    	if (args[i] == "-all") all = true;
-		if (args[i] == "-erase") erase = true;
+    if (args[i] == "-arg") foundArg = true;
+    if (args[i] == "-force") force = true;
+    if (args[i] == "-repo") foundRepo = true;
+    if (args[i] == "-all") all = true;
+		if (args[i] == "-noerase") ferase = false;
 		if (args[i] == "-cred") foundCred = true;
+		if (args[i] == "-erasefolder") derase = false;
 	}
 	var packag = getPackage(args[0]);
 
-    if (all) {
-    	var ops = [];
+	if (all) {
+		var ops = [];
 
-    	if (force) ops.push("-force");
-    	if (erase) ops.push("-erase");
+		if (force) ops.push("-force");
+		if (ferase) ops.push("-noerase");
 
-    	var packages = getOPackLocalDB();
-    	for(let i in packages) {
-    		if (packages[i].name.toUpperCase() == 'OPENAF') continue;
-    		var pack = [];
-    		pack.push(packages[i].name);
-    		pack = pack.concat(ops);
-    		update(pack);
-    	}
+		var packages = getOPackLocalDB();
+		for(let i in packages) {
+			if (packages[i].name.toUpperCase() == 'OPENAF') continue;
+			var pack = [];
+			pack.push(packages[i].name);
+			pack = pack.concat(ops);
+			update(pack);
+		}
 
-    	return;
-    }
+		return;
+	}
 
-	if (!isUndefined(packag) &&
+	if (!isUnDef(packag) &&
 		(typeof packag.name == 'undefined' ||
 		 packag.__filelocation == 'local'))
 		//packag = getOPackRemoteDB()[packag.name];
 		packag = getOPackRemoteDB()[$from(Object.keys(getOPackRemoteDB())).equals(args[0]).select()[0]];
 
 	// Verify version
-	if (!isUndefined(packag) &&
+	if (!isUnDef(packag) &&
 		(typeof packag.name !== 'undefined')) {
 	    if (checkVersion(packag, force)) {
 	    	if ((typeof packag.__filelocation !== 'undefined') &&
@@ -1450,12 +1438,12 @@ function update(args) {
 		return;
 	}
 
-	if (erase) erase(args);
+	if (ferase) erase(clone(args), derase);
 	install(args);
 }
 
 // ERASE
-function erase(args) {
+function erase(args, dontRemoveDir) {
 	if (!isUndefined(args[0]) && args[0].toUpperCase() == 'OPENAF') {
 		logErr("Can't delete OpenAF. Please delete manually or using another package manager (e.g. RPM)");
 		return;
@@ -1526,7 +1514,7 @@ function erase(args) {
 					rmdir(list.files[i].filepath);
 				}
 			}
-			rmdir(args[0]);
+			if (!dontRemoveDir) rmdir(args[0]);
 			log("Package " + packag.name + " erased." + repeat(biggestMessage, " "));
 			removeLocalDB(packag, args[0]);
 			break;
@@ -1749,25 +1737,25 @@ for(let i in verbs) {
 
 		verb = i;
 
-        switch(verb) {
-        case 'info': __opack_info(params); break;
-        case 'install': install(params); break;
-        case 'erase': erase(params); break;
-        case 'list': __opack_list(params); break;
-        case 'genpack': genpack(params); break;
-        case 'pack': pack(params); break;
-        case 'add2db': add(params); break;
-        case 'remove2db': remove(params); break;
-        case 'add2remotedb': addCentral(params); break;
+    switch(verb) {
+		case 'info': __opack_info(params); break;
+		case 'install': install(params); break;
+		case 'erase': erase(params); break;
+		case 'list': __opack_list(params); break;
+		case 'genpack': genpack(params); break;
+		case 'pack': pack(params); break;
+		case 'add2db': add(params); break;
+		case 'remove2db': remove(params); break;
+		case 'add2remotedb': addCentral(params); break;
 		case 'remove2remotedb': removeCentral(params); break;
 		case 'script': __opack_script(params); break;
 		case 'daemon': __opack_script(params, true); break;
 		case 'ojob'  : __opack_script(params, false, true); break;
-        case 'search': __opack_search(params); break;
-        case 'update': update(params); break;
+		case 'search': __opack_search(params); break;
+		case 'update': update(params); break;
 		case 'exec': __opack_exec(params); break;
 		case 'help': showhelp = 1; showHelp(); break;
-        }
+    }
 
 		showhelp = 0;
 	}
